Map Open Fix
Fixed some type casting issues with opening a map from JSON.

+++ core/src/com/TeamHEC/LocomotionCommotion/Map/MapInstance.java
@@ -100,8 +100,8 @@ public class MapInstance {
 				}
 				
 				Long rent = (Long) jStation.get("Rent");
-				Double xPos = (Double) ((JSONArray) jStation.get("Location")).get(0);
-				Double yPos = (Double) ((JSONArray) jStation.get("Location")).get(1);
+				Double xPos = ((Long)((JSONArray) jStation.get("Location")).get(0)).doubleValue();
+				Double yPos = ((Long)((JSONArray) jStation.get("Location")).get(1)).doubleValue();
 				boolean locked = (Boolean) jStation.get("Locked");
 				
 				stations.put(name, new Station(name, baseValue.intValue(), resource, baseFuelOut.intValue(), lines, rent.intValue(), xPos.floatValue(), yPos.floatValue(), locked));
@@ -111,8 +111,8 @@ public class MapInstance {
 			for(int i=0; i<jJunctions.size(); i++) {
 				JSONObject jJunction = (JSONObject) jJunctions.get(i);
 				
-				Double xPos = (Double) ((JSONArray) jJunction.get("Location")).get(0);
-				Double yPos = (Double) ((JSONArray) jJunction.get("Location")).get(1);
+				Double xPos = ((Long)((JSONArray) jJunction.get("Location")).get(0)).doubleValue();
+				Double yPos = ((Long)((JSONArray) jJunction.get("Location")).get(1)).doubleValue();
 				String name = (String) jJunction.get("Name");
 				boolean locked = (Boolean) jJunction.get("Locked");
